refactor(economy): Move schemas to seperate file
